+++ bindings/nodejs/lib/types/client/nodeInfo.ts
@@ -1,7 +1,7 @@
 // Copyright 2021-2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import { InfoResponse } from '../models/api/info';
+import { InfoResponse } from '../models/api/node/info-response';
 
 /** NodeInfo wrapper which contains the node info and the url from the node (useful when multiple nodes are used) */
 export interface NodeInfoResponse {
