+++ bindings/nodejs/lib/types/models/api/blocks/index.ts
@@ -1,0 +1,5 @@
+// Copyright 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './block-response';
+export * from './issuance-response';
