+++ bindings/nodejs/lib/types/models/api/commitments/utxo-changes-response.ts
@@ -1,9 +1,9 @@
 // Copyright 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import { OutputId } from '../../block/output';
-import { SlotCommitmentId } from '../../block';
-import { Output, OutputDiscriminator } from '../../block/output';
+import { OutputId } from '../../../block/output';
+import { SlotCommitmentId } from '../../../block';
+import { Output, OutputDiscriminator } from '../../../block/output';
 import { Type } from 'class-transformer';
 
 /**
