+++ bindings/nodejs/lib/types/models/api/committee/committee-response.ts
@@ -1,8 +1,8 @@
 // Copyright 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import { Bech32Address, EpochIndex } from '../../block';
-import { NumericString } from '../../utils';
+import { Bech32Address, EpochIndex } from '../../../block';
+import { NumericString } from '../../../utils';
 
 /**
  * Returns information of a validator (committee member).
