+++ bindings/nodejs/lib/types/models/api/committee/index.ts
@@ -1,0 +1,4 @@
+// Copyright 2024 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './committee-response';
