+++ bindings/nodejs/lib/types/models/api/index.ts
@@ -1,20 +1,13 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-export * from './info';
+export * from './accounts';
+export * from './blocks';
+export * from './commitments';
+export * from './committee';
+export * from './node';
 export * from './plugins';
-export * from './block-response';
-export * from './committee-response';
-export * from './congestion-response';
-export * from './issuance-response';
-export * from './mana-rewards-response';
-export * from './network-metrics';
-export * from './output-id-proof';
-export * from './output-metadata-response';
-export * from './output-response';
+export * from './rewards';
+export * from './utxo';
 export * from './response';
-export * from './routes-response';
-export * from './transaction-failure-reason';
-export * from './transaction-response';
-export * from './utxo-changes-response';
-export * from './validators-response';
+export * from './validators';
