--- bindings/nodejs/lib/types/models/api/info/index.ts
@@ -1,7 +1,0 @@
-// Copyright 2023 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-export * from './node-info';
-export * from './node-info-base-token';
-export * from './node-info-protocol';
-export * from './node-info-status';
