+++ bindings/nodejs/lib/types/models/api/node/index.ts
@@ -1,0 +1,6 @@
+// Copyright 2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './info-response';
+export * from './network-metrics-response';
+export * from './routes-response';
