+++ bindings/nodejs/lib/types/models/api/node/network-metrics-response.ts
