+++ bindings/nodejs/lib/types/models/api/utxo/output-response.ts
@@ -1,7 +1,7 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import { Output, OutputDiscriminator } from '../../block/output';
+import { Output, OutputDiscriminator } from '../../../block/output';
 import { OutputIdProof } from './output-id-proof';
 
 import { Type } from 'class-transformer';
