+++ bindings/nodejs/lib/types/models/api/utxo/transaction-failure-reason.ts
